+++ Solution/Entity.h
@@ -21,7 +21,8 @@ public:
 	enum ComponentType {
 		SpriteRenderer,
 		RectangleRenderer,
-		EllipseRenderer
+		EllipseRenderer,
+		LineRenderer
 	} ;
 	// whether or not the entity is active in scene. Dictates whether components are called each frame
 	bool isActive = true;