+++ Solution/SpriteRenderer.cpp
@@ -20,7 +20,7 @@ SpriteRenderer::SpriteRenderer(Texture2D* texture, glm::vec3 color, ShaderProgra
 	// set texture
 	this->texture = texture;
 	// if the image texture uses alpha channel then set component property to use transparency
-	if (texture->imageFormat = GL_RGBA)
+	if (texture->imageFormat == GL_RGBA)
 		this->hasTransprency = true;
 	// set colour
 	this->color = color;
@@ -82,7 +82,7 @@ void SpriteRenderer::Draw(std::shared_ptr<OrthoCamera> camera)
 	shaderProgram->SetMatrix4("projection", projection);
 
 	// sey sprite transform
-	shaderProgram->SetMatrix4("modelTransform", spriteTransform.ToMatrix());
+	shaderProgram->SetMatrix4("modelTransform", spriteTransform.ToMatrix(camera));
 	// set color of sprite with alpha channel included
 	shaderProgram->SetVector4f("spriteColor", glm::vec4(color, _alpha));